+++ pyproject.toml
@@ -19,11 +19,6 @@ repository = "https://github.com/golemfactory/golem-core-python"
 [tool.poetry.dependencies]
 python = "^3.8.0"
 
-#   TODO: We use quite a lot of yapapi dependencies that are not implicitly imported,
-#         this should be fixed - but not now. First we must be sure what is the future of
-#         golem_core, yapapi and how they are related.
-yapapi = "^0.9.3"
-
 #   CLI
 prettytable = "^3.4.1"
 click = "^8.1.3"
